+++ dispike/__init__.py
@@ -1,7 +1,7 @@
 __version__ = "1.0.1-beta.0"
 
 from .main import Dispike
-from .models import IncomingDiscordInteraction, IncomingApplicationCommand
+from .incoming import IncomingDiscordInteraction, IncomingApplicationCommand
 from .response import DiscordResponse
-from .register.models import DiscordCommand
+from .creating.models import DiscordCommand
 from .interactions import PerCommandRegistrationSettings, EventCollection
