--- dispike/models/__init__.py
@@ -1,8 +1,0 @@
-from .discord_types.member import Member
-from .discord_types.user import User
-from .incoming import (
-    IncomingDiscordInteraction,
-    IncomingDiscordOption,
-    IncomingDiscordOptionList,
-    IncomingApplicationCommand,
-)
