--- dispike/register/__init__.py
@@ -1,1 +1,0 @@
-from .registrator import RegisterCommands
