+++ tests/test_models.py
@@ -1,7 +1,10 @@
-from dispike.models import *
+from dispike.incoming import *
 import pytest
 
-from dispike.models.incoming import IncomingDiscordButtonInteraction, IncomingDiscordSelectMenuInteraction
+from dispike.incoming.incoming_interactions import (
+    IncomingDiscordButtonInteraction,
+    IncomingDiscordSelectMenuInteraction,
+)
 
 
 def test_valid_incoming_one_option():
@@ -49,10 +52,7 @@ def test_valid_incoming_one_option():
 def test_valid_incoming_button():
     data = {
         "channel_id": "123123",
-        "data": {
-            "custom_id": "click_one",
-            "component_type": 2
-        },
+        "data": {"custom_id": "click_one", "component_type": 2},
         "guild_id": "123123",
         "id": "123123123132",
         "member": {
@@ -94,10 +94,7 @@ def test_valid_incoming_select_menu():
         "data": {
             "component_type": 3,
             "custom_id": "class_select_1",
-            "values": [
-                "mage",
-                "rogue"
-            ]
+            "values": ["mage", "rogue"],
         },
         "guild_id": "123123",
         "id": "123123123132",
@@ -132,10 +129,7 @@ def test_valid_incoming_select_menu():
 
     _create_model = IncomingDiscordSelectMenuInteraction(**data)
     assert _create_model.data.custom_id == "class_select_1"
-    assert _create_model.data.values == [
-        "mage",
-        "rogue"
-    ]
+    assert _create_model.data.values == ["mage", "rogue"]
 
 
 def test_valid_incoming_multiple_options():
